Navigation on a CardClick added
